+++ development/ui/dashboard_ui/educational_new_theme/client-side/src/app/reports/attendance/student-attendance/student-attendance.component.html
@@ -1,13 +1,14 @@
 <div class="container-fluid">
     <div class="row reportHeader">
         <div class="col-lg-12 col-md-12 col-sm-12">
-            <div class="row" (window:resize)="commonService.onResize(levelWise);onResize()">
+            <div class="row" (window:resize)="globalService.onResize(levelWise);onResize()">
                 <!-- Level wise hierarchy  ===========-->
                 <div class="col-lg-7 col-md-12 col-sm-12 ">
                     <div *ngIf='skul'>
                         <p id="dist">
                             Student Attendance Report for: <span id="state" (click)="onClickHome()"
-                                style="cursor: pointer;text-decoration: underline;" class="selected-level">{{state}}</span></p>
+                                style="cursor: pointer;text-decoration: underline;"
+                                class="selected-level">{{state}}</span></p>
                     </div>
                     <div *ngIf='dist'>
                         <p id="block">
