+++ development/ui/dashboard_ui/educational_new_theme/client-side/src/app/reports/attendance/student-attendance/student-attendance.component.ts
@@ -11,7 +11,8 @@ import { Router } from "@angular/router";
 import * as L from "leaflet";
 import * as R from "leaflet-responsive-popup";
 import { KeycloakSecurityService } from "../../../keycloak-security.service";
-import { AppServiceComponent, globalMap } from "../../../app.service";
+import { AppServiceComponent } from "../../../app.service";
+import { MapService, globalMap } from "src/app/services/map-services/maps.service";
 declare const $;
 
 @Component({
@@ -108,6 +109,8 @@ export class StudengtAttendanceComponent implements OnInit {
 
   selected = "absolute";
   reportName = "student_attendance";
+  mapName;
+  // globalService;
 
   getColor(data) {
     this.selected = data;
@@ -135,7 +138,8 @@ export class StudengtAttendanceComponent implements OnInit {
     public keyCloakSevice: KeycloakSecurityService,
     private changeDetection: ChangeDetectorRef,
     public commonService: AppServiceComponent,
-    private readonly _router: Router
+    private readonly _router: Router,
+    public globalService: MapService,
   ) {
     this.commonService.callProgressCard.subscribe(value => {
       if (value) {
@@ -161,11 +165,12 @@ export class StudengtAttendanceComponent implements OnInit {
   public allMonths: any = ['June', 'July', 'August', 'September', 'October', 'November', 'December', 'January', 'February', 'March', 'April', 'May'];
 
   ngOnInit() {
+    this.mapName = this.commonService.mapName;
     this.state = this.commonService.state;
-    this.commonService.latitude = this.lat = this.commonService.mapCenterLatlng.lat;
-    this.commonService.longitude = this.lng = this.commonService.mapCenterLatlng.lng;
+    this.globalService.latitude = this.lat = this.globalService.mapCenterLatlng.lat;
+    this.globalService.longitude = this.lng = this.globalService.mapCenterLatlng.lng;
     this.changeDetection.detectChanges();
-    this.commonService.initMap("sarMap", [[this.lat, this.lng]]);
+    this.globalService.initMap("sarMap", [[this.lat, this.lng]]);
     document.getElementById("accessProgressCard").style.display = "block";
     document.getElementById("backBtn") ? document.getElementById("backBtn").style.display = "none" : "";
     this.skul = true;
@@ -655,9 +660,8 @@ export class StudengtAttendanceComponent implements OnInit {
                     id: this.markers[i]["district_id"],
                     name: this.markers[i]["district_name"],
                   });
-
                   //initialize markers with its latitude and longitude
-                  var markerIcon = this.commonService.initMarkers1(
+                  var markerIcon = this.globalService.initMarkers1(
                     this.markers[i].lat,
                     this.markers[i].lng,
                     this.selected == "absolute"
@@ -688,8 +692,7 @@ export class StudengtAttendanceComponent implements OnInit {
                 a.name > b.name ? 1 : b.name > a.name ? -1 : 0
               );
               this.districtsNames = distNames;
-
-              this.commonService.restrictZoom(globalMap);
+              this.globalService.restrictZoom(globalMap);
 
               //Setting map bound for scroll::::::::::::
               globalMap.setMaxBounds([
@@ -698,7 +701,7 @@ export class StudengtAttendanceComponent implements OnInit {
               ]);
 
               //adjusting marker size and other UI on screen resize:::::::::::
-              this.commonService.onResize(this.levelWise);
+              this.globalService.onResize(this.levelWise);
               this.commonService.loaderAndErr(this.markers);
               this.changeDetection.markForCheck();
             },
@@ -782,7 +785,7 @@ export class StudengtAttendanceComponent implements OnInit {
                     distId: this.markers[i]["dist"],
                   });
                   //initialize markers with its latitude and longitude
-                  var markerIcon = this.commonService.initMarkers1(
+                  var markerIcon = this.globalService.initMarkers1(
                     this.markers[i].lat,
                     this.markers[i].lng,
                     this.selected == "absolute"
@@ -812,7 +815,7 @@ export class StudengtAttendanceComponent implements OnInit {
               );
               this.blocksNames = blockNames;
 
-              this.commonService.restrictZoom(globalMap);
+              this.globalService.restrictZoom(globalMap);
 
               //Setting map bound for scroll::::::::::::
               globalMap.setMaxBounds([
@@ -821,7 +824,7 @@ export class StudengtAttendanceComponent implements OnInit {
               ]);
 
               //adjusting marker size and other UI on screen resize:::::::::::
-              this.commonService.onResize(this.levelWise);
+              this.globalService.onResize(this.levelWise);
               this.commonService.loaderAndErr(this.markers);
               this.changeDetection.markForCheck();
             },
@@ -920,7 +923,7 @@ export class StudengtAttendanceComponent implements OnInit {
                     distId: this.markers[i]["district_id"],
                   });
                   //initialize markers with its latitude and longitude
-                  var markerIcon = this.commonService.initMarkers1(
+                  var markerIcon = this.globalService.initMarkers1(
                     this.markers[i].lat,
                     this.markers[i].lng,
                     this.selected == "absolute"
@@ -955,7 +958,7 @@ export class StudengtAttendanceComponent implements OnInit {
               );
               this.blocksNames = blockNames;
 
-              this.commonService.restrictZoom(globalMap);
+              this.globalService.restrictZoom(globalMap);
 
               //Setting map bound for scroll::::::::::::
               globalMap.setMaxBounds([
@@ -964,7 +967,7 @@ export class StudengtAttendanceComponent implements OnInit {
               ]);
 
               //adjusting marker size and other UI on screen resize:::::::::::
-              this.commonService.onResize(this.levelWise);
+              this.globalService.onResize(this.levelWise);
               this.commonService.loaderAndErr(this.markers);
               this.changeDetection.markForCheck();
             },
@@ -1043,7 +1046,7 @@ export class StudengtAttendanceComponent implements OnInit {
                   );
                   this.districtsIds.push(sorted[i]["district_id"]);
                   //initialize markers with its latitude and longitude
-                  var markerIcon = this.commonService.initMarkers1(
+                  var markerIcon = this.globalService.initMarkers1(
                     this.markers[i].lat,
                     this.markers[i].lng,
                     this.selected == "absolute"
@@ -1080,7 +1083,7 @@ export class StudengtAttendanceComponent implements OnInit {
               ]);
 
               //adjusting marker size and other UI on screen resize:::::::::::
-              this.commonService.onResize(this.levelWise);
+              this.globalService.onResize(this.levelWise);
               this.commonService.loaderAndErr(this.markers);
               this.changeDetection.markForCheck();
             },
@@ -1124,8 +1127,8 @@ export class StudengtAttendanceComponent implements OnInit {
     this.title = "";
     this.titleName = "";
     this.clustName = "";
-    this.commonService.latitude = this.lat = this.commonService.mapCenterLatlng.lat;
-    this.commonService.longitude = this.lng = this.commonService.mapCenterLatlng.lng;
+    this.globalService.latitude = this.lat = this.globalService.mapCenterLatlng.lat;
+    this.globalService.longitude = this.lng = this.globalService.mapCenterLatlng.lng;
 
     //Setting map bound for scroll::::::::::::
     globalMap.setMaxBounds([
@@ -1328,10 +1331,10 @@ export class StudengtAttendanceComponent implements OnInit {
               },
                 []);
               this.mylatlngData = uniqueData;
-              this.commonService.latitude = this.lat = Number(
+              this.globalService.latitude = this.lat = Number(
                 this.mylatlngData[0]["lat"]
               );
-              this.commonService.longitude = this.lng = Number(
+              this.globalService.longitude = this.lng = Number(
                 this.mylatlngData[0]["lng"]
               );
 
@@ -1362,7 +1365,7 @@ export class StudengtAttendanceComponent implements OnInit {
                     name: this.markers[i]["block_name"],
                   });
                   //initialize markers with its latitude and longitude
-                  var markerIcon = this.commonService.initMarkers1(
+                  var markerIcon = this.globalService.initMarkers1(
                     this.markers[i].lat,
                     this.markers[i].lng,
                     this.selected == "absolute"
@@ -1392,7 +1395,7 @@ export class StudengtAttendanceComponent implements OnInit {
               );
               this.blocksNames = blokName;
 
-              this.commonService.restrictZoom(globalMap);
+              this.globalService.restrictZoom(globalMap);
 
               //Setting map bound for scroll::::::::::::
               globalMap.setMaxBounds([
@@ -1401,7 +1404,7 @@ export class StudengtAttendanceComponent implements OnInit {
               ]);
 
               //adjusting marker size and other UI on screen resize:::::::::::
-              this.commonService.onResize(this.levelWise);
+              this.globalService.onResize(this.levelWise);
               this.commonService.loaderAndErr(this.markers);
               this.changeDetection.markForCheck();
             },
@@ -1534,10 +1537,10 @@ export class StudengtAttendanceComponent implements OnInit {
               },
                 []);
               this.mylatlngData = uniqueData;
-              this.commonService.latitude = this.lat = Number(
+              this.globalService.latitude = this.lat = Number(
                 this.mylatlngData[0]["lat"]
               );
-              this.commonService.longitude = this.lng = Number(
+              this.globalService.longitude = this.lng = Number(
                 this.mylatlngData[0]["lng"]
               );
               var clustNames = [];
@@ -1579,7 +1582,7 @@ export class StudengtAttendanceComponent implements OnInit {
                     });
                   }
                   //initialize markers with its latitude and longitude
-                  var markerIcon = this.commonService.initMarkers1(
+                  var markerIcon = this.globalService.initMarkers1(
                     this.markers[i].lat,
                     this.markers[i].lng,
                     this.selected == "absolute"
@@ -1610,7 +1613,7 @@ export class StudengtAttendanceComponent implements OnInit {
               );
               this.clusterNames = clustNames;
 
-              this.commonService.restrictZoom(globalMap);
+              this.globalService.restrictZoom(globalMap);
 
               //Setting map bound for scroll::::::::::::
               globalMap.setMaxBounds([
@@ -1619,7 +1622,7 @@ export class StudengtAttendanceComponent implements OnInit {
               ]);
 
               //adjusting marker size and other UI on screen resize:::::::::::
-              this.commonService.onResize(this.levelWise);
+              this.globalService.onResize(this.levelWise);
               this.commonService.loaderAndErr(this.markers);
               this.changeDetection.markForCheck();
             },
@@ -1790,10 +1793,10 @@ export class StudengtAttendanceComponent implements OnInit {
               },
                 []);
               this.mylatlngData = uniqueData;
-              this.commonService.latitude = this.lat = Number(
+              this.globalService.latitude = this.lat = Number(
                 this.mylatlngData[0]["lat"]
               );
-              this.commonService.longitude = this.lng = Number(
+              this.globalService.longitude = this.lng = Number(
                 this.mylatlngData[0]["lng"]
               );
 
@@ -1819,7 +1822,7 @@ export class StudengtAttendanceComponent implements OnInit {
                     "attendance"
                   );
                   //initialize markers with its latitude and longitude
-                  var markerIcon = this.commonService.initMarkers1(
+                  var markerIcon = this.globalService.initMarkers1(
                     this.markers[i].lat,
                     this.markers[i].lng,
                     this.selected == "absolute"
@@ -1854,7 +1857,7 @@ export class StudengtAttendanceComponent implements OnInit {
               ]);
 
               //adjusting marker size and other UI on screen resize:::::::::::
-              this.commonService.onResize(this.levelWise);
+              this.globalService.onResize(this.levelWise);
               this.commonService.loaderAndErr(this.markers);
               this.changeDetection.markForCheck();
             },
@@ -1913,15 +1916,14 @@ export class StudengtAttendanceComponent implements OnInit {
       }
     });
 
-    var yourData = this.commonService
-      .getInfoFrom(
-        orgObject,
-        "attendance",
-        levelWise,
-        "std-attd",
-        undefined,
-        undefined
-      )
+    var yourData = this.globalService.getInfoFrom(
+      orgObject,
+      "attendance",
+      levelWise,
+      "std-attd",
+      undefined,
+      undefined
+    )
       .join(" <br>");
     const popup = R.responsivePopup({
       hasTip: false,
@@ -2129,7 +2131,7 @@ export class StudengtAttendanceComponent implements OnInit {
         this.studentCount += parseInt(markers[i]['number_of_students'].replace(',', ''));
 
         //initialize markers with its latitude and longitude
-        var markerIcon = this.commonService.initMarkers1(
+        var markerIcon = this.globalService.initMarkers1(
           markers[i].lat,
           markers[i].lng,
           this.selected == "absolute"
@@ -2183,7 +2185,7 @@ export class StudengtAttendanceComponent implements OnInit {
     }
 
     //adjusting marker size and other UI on screen resize:::::::::::
-    this.commonService.onResize(this.levelWise);
+    this.globalService.onResize(this.levelWise);
     this.commonService.loaderAndErr(markers)
     this.changeDetection.detectChanges();
   }
