+++ development/ui/dashboard_ui/educational_new_theme/client-side/src/app/reports/exception-list/teacher-attendance-exception/teacher-attendance-exception.component.ts
@@ -11,7 +11,8 @@ import { Router } from "@angular/router";
 import * as L from "leaflet";
 import * as R from "leaflet-responsive-popup";
 import { KeycloakSecurityService } from "../../../keycloak-security.service";
-import { AppServiceComponent, globalMap } from "../../../app.service";
+import { AppServiceComponent } from "../../../app.service";
+import { MapService, globalMap } from "src/app/services/map-services/maps.service";
 declare const $;
 
 @Component({
@@ -105,7 +106,8 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
     public keyCloakSevice: KeycloakSecurityService,
     private changeDetection: ChangeDetectorRef,
     public commonService: AppServiceComponent,
-    private readonly _router: Router
+    private readonly _router: Router,
+    public globalService: MapService,
   ) { }
 
   width = window.innerWidth;
@@ -117,10 +119,10 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
 
   ngOnInit() {
     this.state = this.commonService.state;
-    this.commonService.latitude = this.lat = this.commonService.mapCenterLatlng.lat;
-    this.commonService.longitude = this.lng = this.commonService.mapCenterLatlng.lng;
+    this.globalService.latitude = this.lat = this.globalService.mapCenterLatlng.lat;
+    this.globalService.longitude = this.lng = this.globalService.mapCenterLatlng.lng;
     this.changeDetection.detectChanges();
-    this.commonService.initMap("tarExpMap", [[this.lat, this.lng]]);
+    this.globalService.initMap("tarExpMap", [[this.lat, this.lng]]);
     globalMap.setMaxBounds([
       [this.lat - 4.5, this.lng - 6],
       [this.lat + 3.5, this.lng + 6],
@@ -405,7 +407,7 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
                   id: this.markers[i]["district_id"],
                   name: this.markers[i]["district_name"],
                 });
-                var markerIcon = this.commonService.initMarkers1(
+                var markerIcon = this.globalService.initMarkers1(
                   this.markers[i].lat,
                   this.markers[i].lng,
                   this.commonService.relativeColorGredient(
@@ -435,12 +437,12 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
             );
             this.districtsNames = distNames;
 
-            this.commonService.restrictZoom(globalMap);
+            this.globalService.restrictZoom(globalMap);
             globalMap.setMaxBounds([
               [this.lat - 4.5, this.lng - 6],
               [this.lat + 3.5, this.lng + 6],
             ]);
-            this.commonService.onResize(this.level);
+            this.globalService.onResize(this.level);
             this.schoolCount = this.schoolCount
               .toString()
               .replace(/(\d)(?=(\d\d)+\d$)/g, "$1,");
@@ -507,7 +509,7 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
                   name: this.markers[i]["block_name"],
                   distId: this.markers[i]["dist"],
                 });
-                var markerIcon = this.commonService.initMarkers1(
+                var markerIcon = this.globalService.initMarkers1(
                   this.markers[i].lat,
                   this.markers[i].lng,
                   this.commonService.relativeColorGredient(
@@ -535,12 +537,12 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
               );
               this.blocksNames = blockNames;
 
-              this.commonService.restrictZoom(globalMap);
+              this.globalService.restrictZoom(globalMap);
               globalMap.setMaxBounds([
                 [this.lat - 4.5, this.lng - 6],
                 [this.lat + 3.5, this.lng + 6],
               ]);
-              this.commonService.onResize(this.level);
+              this.globalService.onResize(this.level);
               this.schoolCount = this.schoolCount
                 .toString()
                 .replace(/(\d)(?=(\d\d)+\d$)/g, "$1,");
@@ -624,7 +626,7 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
                   name: this.markers[i]["block_name"],
                   distId: this.markers[i]["district_id"],
                 });
-                var markerIcon = this.commonService.initMarkers1(
+                var markerIcon = this.globalService.initMarkers1(
                   this.markers[i].lat,
                   this.markers[i].lng,
                   this.commonService.relativeColorGredient(
@@ -657,12 +659,12 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
               );
               this.blocksNames = blockNames;
 
-              this.commonService.restrictZoom(globalMap);
+              this.globalService.restrictZoom(globalMap);
               globalMap.setMaxBounds([
                 [this.lat - 4.5, this.lng - 6],
                 [this.lat + 3.5, this.lng + 6],
               ]);
-              this.commonService.onResize(this.level);
+              this.globalService.onResize(this.level);
               this.schoolCount = this.schoolCount
                 .toString()
                 .replace(/(\d)(?=(\d\d)+\d$)/g, "$1,");
@@ -723,7 +725,7 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
             if (this.markers.length !== 0) {
               for (let i = 0; i < this.markers.length; i++) {
                 this.districtsIds.push(sorted[i]["district_id"]);
-                var markerIcon = this.commonService.initMarkers1(
+                var markerIcon = this.globalService.initMarkers1(
                   this.markers[i].lat,
                   this.markers[i].lng,
                   "red",
@@ -746,7 +748,7 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
                 [this.lat - 4.5, this.lng - 6],
                 [this.lat + 3.5, this.lng + 6],
               ]);
-              this.commonService.onResize(this.level);
+              this.globalService.onResize(this.level);
               this.schoolCount = this.markers.length
                 .toString()
                 .replace(/(\d)(?=(\d\d)+\d$)/g, "$1,");
@@ -792,8 +794,8 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
     this.title = "";
     this.titleName = "";
     this.clustName = "";
-    this.commonService.latitude = this.lat = this.commonService.mapCenterLatlng.lat;
-    this.commonService.longitude = this.lng = this.commonService.mapCenterLatlng.lng;
+    this.globalService.latitude = this.lat = this.globalService.mapCenterLatlng.lat;
+    this.globalService.longitude = this.lng = this.globalService.mapCenterLatlng.lng;
     globalMap.setMaxBounds([
       [this.lat - 4.5, this.lng - 6],
       [this.lat + 3.5, this.lng + 6],
@@ -981,10 +983,10 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
             },
               []);
             this.mylatlngData = uniqueData;
-            this.commonService.latitude = this.lat = Number(
+            this.globalService.latitude = this.lat = Number(
               this.mylatlngData[0]["lat"]
             );
-            this.commonService.longitude = this.lng = Number(
+            this.globalService.longitude = this.lng = Number(
               this.mylatlngData[0]["lng"]
             );
 
@@ -1004,7 +1006,7 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
                 id: this.markers[i]["block_id"],
                 name: this.markers[i]["block_name"],
               });
-              var markerIcon = this.commonService.initMarkers1(
+              var markerIcon = this.globalService.initMarkers1(
                 this.markers[i].lat,
                 this.markers[i].lng,
                 this.commonService.relativeColorGredient(
@@ -1032,12 +1034,12 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
             );
             this.blocksNames = blokName;
 
-            this.commonService.restrictZoom(globalMap);
+            this.globalService.restrictZoom(globalMap);
             globalMap.setMaxBounds([
               [this.lat - 1.5, this.lng - 3],
               [this.lat + 1.5, this.lng + 2],
             ]);
-            this.commonService.onResize(this.level);
+            this.globalService.onResize(this.level);
             this.schoolCount = this.schoolCount
               .toString()
               .replace(/(\d)(?=(\d\d)+\d$)/g, "$1,");
@@ -1161,10 +1163,10 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
             },
               []);
             this.mylatlngData = uniqueData;
-            this.commonService.latitude = this.lat = Number(
+            this.globalService.latitude = this.lat = Number(
               this.mylatlngData[0]["lat"]
             );
-            this.commonService.longitude = this.lng = Number(
+            this.globalService.longitude = this.lng = Number(
               this.mylatlngData[0]["lng"]
             );
             var clustNames = [];
@@ -1194,7 +1196,7 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
                   blockId: sorted[i]["block_id"],
                 });
               }
-              var markerIcon = this.commonService.initMarkers1(
+              var markerIcon = this.globalService.initMarkers1(
                 this.markers[i].lat,
                 this.markers[i].lng,
                 this.commonService.relativeColorGredient(
@@ -1223,12 +1225,12 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
             );
             this.clusterNames = clustNames;
 
-            this.commonService.restrictZoom(globalMap);
+            this.globalService.restrictZoom(globalMap);
             globalMap.setMaxBounds([
               [this.lat - 1.5, this.lng - 3],
               [this.lat + 1.5, this.lng + 2],
             ]);
-            this.commonService.onResize(this.level);
+            this.globalService.onResize(this.level);
             this.schoolCount = this.schoolCount
               .toString()
               .replace(/(\d)(?=(\d\d)+\d$)/g, "$1,");
@@ -1388,10 +1390,10 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
             },
               []);
             this.mylatlngData = uniqueData;
-            this.commonService.latitude = this.lat = Number(
+            this.globalService.latitude = this.lat = Number(
               this.mylatlngData[0]["lat"]
             );
-            this.commonService.longitude = this.lng = Number(
+            this.globalService.longitude = this.lng = Number(
               this.mylatlngData[0]["lng"]
             );
 
@@ -1406,7 +1408,7 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
             });
             this.markers = sorted;
             for (var i = 0; i < sorted.length; i++) {
-              var markerIcon = this.commonService.initMarkers1(
+              var markerIcon = this.globalService.initMarkers1(
                 this.markers[i].lat,
                 this.markers[i].lng,
                 this.commonService.relativeColorGredient(
@@ -1435,7 +1437,7 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
               [this.lat - 1.5, this.lng - 3],
               [this.lat + 1.5, this.lng + 2],
             ]);
-            this.commonService.onResize(this.level);
+            this.globalService.onResize(this.level);
             this.schoolCount = this.markers.length
               .toString()
               .replace(/(\d)(?=(\d\d)+\d$)/g, "$1,");
@@ -1499,15 +1501,14 @@ export class TeacherAttendanceExceptionComponent implements OnInit {
       }
     });
 
-    var yourData = this.commonService
-      .getInfoFrom(
-        orgObject,
-        "attendance",
-        levelWise,
-        "std-attd",
-        undefined,
-        undefined
-      )
+    var yourData = this.globalService.getInfoFrom(
+      orgObject,
+      "attendance",
+      levelWise,
+      "std-attd",
+      undefined,
+      undefined
+    )
       .join(" <br>");
     const popup = R.responsivePopup({
       hasTip: false,
