+++ development/ui/dashboard_ui/educational_new_theme/client-side/src/app/reports/telemetry/telemetry-data/telemetry-data.component.html
@@ -1,10 +1,7 @@
 <div class="container-fluid">
   <div class="row reportHeader">
     <div class="col-lg-12 col-md-12 col-sm-12">
-      <div
-        class="row"
-        (window:resize)="commonService.onResize(level); onResize()"
-      >
+      <div class="row" (window:resize)="globalService.onResize(level); onResize()">
         <div class="col-sm-8">
           <div>
             <p id="dist" *ngIf="skul">Telemetry Data for: District Level</p>
@@ -66,28 +63,16 @@
             <div class="col-sm-7">
               <div class="row">
                 <div class="col-sm-4">
-                  <em
-                    *ngIf="level != 'District'"
-                    class="fa fa-home"
-                    id="home"
-                    title="Home"
-                    alt="Home"
-                    (click)="districtWise()"
-                    style="
+                  <em *ngIf="level != 'District'" class="fa fa-home" id="home" title="Home" alt="Home"
+                    (click)="districtWise()" style="
                       display: block !important;
                       float: right !important;
                       margin-right: 10px;
-                    "
-                  ></em>
+                    "></em>
                 </div>
                 <div class="col-sm-8">
-                  <button
-                    *ngIf="reportData.length > 0"
-                    class="btn btn-primary"
-                    title="Download Report"
-                    id="download1"
-                    (click)="downloadReport()"
-                  >
+                  <button *ngIf="reportData.length > 0" class="btn btn-primary" title="Download Report" id="download1"
+                    (click)="downloadReport()">
                     Download Report &nbsp;
                     <i class="fa fa-arrow-down" aria-hidden="true"></i>
                   </button>
@@ -99,37 +84,21 @@
       </div>
       <div class="row" style="margin-bottom: 10px">
         <div class="col-sm-12">
-          <select
-            id="period"
-            [(ngModel)]="timePeriod"
-            name="timePeriod"
-            (change)="getTimePeriod(timePeriod)"
-          >
+          <select id="period" [(ngModel)]="timePeriod" name="timePeriod" (change)="getTimePeriod(timePeriod)">
             <option [value]="timePeriod" disabled selected="selected">
               Choose Time Period
             </option>
-            <option
-              *ngFor="let time of timeDetails; let i = index"
-              [(ngValue)]="time.id"
-            >
+            <option *ngFor="let time of timeDetails; let i = index" [(ngValue)]="time.id">
               {{ time.time }}
             </option>
           </select>
           <button class="btn btn-secondary" id="blockbtn" (click)="blockWise()">
             &nbsp;<strong>Blocks</strong>
           </button>
-          <button
-            class="btn btn-secondary"
-            id="clusterbtn"
-            (click)="clusterWise()"
-          >
+          <button class="btn btn-secondary" id="clusterbtn" (click)="clusterWise()">
             &nbsp;<strong>Clusters</strong>
           </button>
-          <button
-            class="btn btn-secondary"
-            id="schoolbtn"
-            (click)="schoolWise()"
-          >
+          <button class="btn btn-secondary" id="schoolbtn" (click)="schoolWise()">
             &nbsp;<strong>Schools</strong>
           </button>
         </div>
@@ -138,12 +107,7 @@
   </div>
   <div class="row reportContainer1">
     <div class="col-sm-12">
-      <div
-        #mapContainer
-        id="map"
-        class="active"
-        style="width: 100%"
-        [ngStyle]="{
+      <div #mapContainer id="map" class="active" style="width: 100%" [ngStyle]="{
           height:
             height > 1760
               ? '88vh'
@@ -152,8 +116,7 @@
               : height > 667 && height < 1160
               ? '80vh'
               : '74vh'
-        }"
-      ></div>
+        }"></div>
     </div>
   </div>
 </div>
