+++ development/ui/dashboard_ui/educational_new_theme/client-side/src/app/services/map-services/google-map.service.ts
@@ -1,0 +1,9 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class GoogleMapService {
+
+  constructor() { }
+}
