+++ development/ui/dashboard_ui/educational_new_theme/client-side/src/environments/environment.ts
@@ -13,7 +13,8 @@ export const environment = {
   stateName: "GJ",
   useCase: "education_usecase",
   diksha_columns: false,
-  theme: "theme1"
+  theme: "theme1",
+  mapName: "mapmyindia"
 };
 
 /*
