+++ development/ui/dashboard_ui/educational_new_theme/client-side/src/index.html
@@ -6,6 +6,10 @@
   <title>cQube</title>
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="icon" type="image/x-icon" href="./assets/images/icon.ico">
+
+  <script src="https://apis.mapmyindia.com/advancedmaps/v1/5068cc8faec963dbaaef34518ab9cbb1/map_load?v=1.5">
+  </script>
+
 </head>
 
 <body>
